requirements.txt
django==4.2
matplotlib==3.10.8
numpy==2.4.2
ollama==0.6.1
Pillow==12.1.0
scikit_learn==1.8.0
seaborn==0.13.2
torch==2.9.1
torchvision==0.24.1
gunicorn
whitenoise
